requirements.txt
python-dotenv >= 1.0.1
requests >= 2.32.3
pillow >= 10.3.0
shapely >= 2.0.4
torch >= 2.3.0
torchvision >= 0.18.0
vit-pytorch >= 1.6.9
matplotlib >= 3.9.0